third_party/WebKit/LayoutTests/crypto/subtle/ecdsa/sign-verify-expected.txt

Test importing an EC key for ECDSA.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Importing a public key...

Importing a private key...

Signing some text...

Verifying the signature...
PASS verificationResult is true

Verifying a bad signature...
PASS verificationResult is false
PASS successfullyParsed is true

TEST COMPLETE
